Move codebase into src
