+++ src/tests/scenarios/xpathscenario_test.py
@@ -1,6 +1,6 @@
 import unittest
 import httpretty
-from src.scenarios.xpathscenario import XPathScenario
+from scenarios.xpathscenario import XPathScenario
 
 SAMPLE_BODY = r"""
 <html>
